fixed test configuration
