+++ package.json
@@ -9,7 +9,7 @@
   "scripts": {
     "console": "node console.js",
     "cover": "NODE_ENV=testing jest --coverage --runInBand --forceExit test/app",
-    "test": "NODE_ENV=testing jest test/app --runInBand --forceExit",
+    "test": "NODE_ENV=testing jest test/app --runInBand --forceExit --passWithNoTests --detectOpenHandles",
     "test-inspect": "NODE_ENV=testing node --inspect --debug-brk jest test/app.spec.js",
     "eslint-check": "eslint --print-config .eslintrc.js --ignore-pattern ./.eslintrc.js | eslint-config-prettier-check",
     "lint": "eslint \"**/*.js\" --ignore-pattern ./.eslintrc.js",
@@ -49,10 +49,7 @@
       "!**/scripts/**",
       "!**/docs/**"
     ],
-    "testEnvironment": "node",
-    "transform": {
-      "^.+\\.js$": "babel-jest"
-    }
+    "setupFilesAfterEnv": ["<rootDir>/test/setup.js"]
   },
   "dependencies": {
     "bcryptjs": "^2.4.3",
