--- test/app.spec.js
@@ -1,31 +1,0 @@
-'use strict';
-
-const fs = require('fs'),
-  models = require('../app/models'),
-  path = require('path');
-
-const tables = Object.values(models.sequelize.models);
-
-const truncateTable = model =>
-  model.destroy({ truncate: true, cascade: true, force: true, restartIdentity: true });
-
-const truncateDatabase = () => Promise.all(tables.map(truncateTable));
-
-beforeEach(done => {
-  truncateDatabase().then(() => done());
-});
-
-// including all test files
-const normalizedPath = path.join(__dirname, '.');
-
-const requireAllTestFiles = pathToSearch => {
-  fs.readdirSync(pathToSearch).forEach(file => {
-    if (fs.lstatSync(`${pathToSearch}/${file}`).isDirectory()) {
-      requireAllTestFiles(`${pathToSearch}/${file}`);
-    } else {
-      require(`${pathToSearch}/${file}`);
-    }
-  });
-};
-
-requireAllTestFiles(normalizedPath);
